feat(blog): add markdown to HTML conversion function & frontmatter
This commit introduces a new function `convertMarkdownContentToHTML`
that takes a markdown string and converts it to HTML. This function
is used in two handlers: `handleGetBlogPost` and `handleGetBlogModal`.
The conversion logic has been removed from `handleGetBlogPost` and
replaced with a call to the new function. A new route
`GET /posts/content` has been added which uses the new
`handleGetBlogModal` handler.