feat: Add dynamic header image to blog posts
This commit introduces a change where the header image for blog posts
is now dynamic. Previously, the header image was a static icon. Now,
the header image is set based on the `HeaderImage` field of the blog
post. This allows for more customization and better representation of
the blog post content.