refactor(cd.yml): simplify static file sync with GCP
This commit simplifies the process of syncing static files with
Google Cloud Platform in the continuous deployment workflow file
(cd.yml). Instead of checking for changes in ./ui/static/*/* and
uploading only the changed files, the new process syncs the entire
./ui/static directory with the GCP bucket using the 'gcloud storage
rsync' command. This change makes the workflow more efficient and
easier to maintain.