feat: Add ping endpoint and tests
This commit introduces a new ping endpoint to the application. This
endpoint responds with a simple "OK" message and can be used for
health checks or simple connectivity tests.

A new test has been added to verify the functionality of the ping
endpoint. The test ensures that the endpoint responds with a 200
status code and the expected "OK" message.

Additionally, a new test utility has been added to facilitate the
creation of a test application and server. This utility will be
useful for future tests as well.