feat(models): introduce interfaces for mocking models
This commit introduces interfaces for the UserModel and PostModel. This
allows for the creation of mock models for testing purposes. The
mock models are implemented in the mocks package. The changes also
include updates to the application struct to use the new interfaces
instead of the concrete models. This makes the application more
flexible and testable.