feat: Introduce social links and optimize CSS (#16)
* feat: Introduce social links and optimize CSS

This commit introduces social links to the base templates of the
website. The social links are displayed at the bottom of the page
and include LinkedIn, GitHub, and Twitter. The display of these
links is controlled by a new `SocialLinks` component.

The commit also includes some minor CSS changes to improve the
layout and responsiveness of the website. The loading of the
`htmx.min.js` script has been modified to only load for
authenticated users, improving performance for unauthenticated users.

Additionally, the mobile menu has been commented out as part of a
redesign process, and the width of the post page has been adjusted
for better readability.

* update article width and fix possible integer overflow issue